Add todo by Id route and refactor code
- Add try and catch in other route
- Change todo model
- Add validate object id to route

+++ backend/routes/todos.js
@@ -3,12 +3,41 @@ var router = express.Router();
 const multer = require("multer");
 const upload = multer(); // for parsing multipart/form-data
 var { TodoModel } = require("../model/Todo");
+const mongoose = require("mongoose");
 
-router.get("/", async function (req, res) {
-  const todos = await TodoModel.find();
-  console.log("todos", todos);
-  const plainTodos = todos.map((todo) => todo.toObject());
-  res.send({ result: plainTodos });
+const validateObjectId = (objId) => {
+  return mongoose.isValidObjectId(objId);
+};
+
+router.get("/", async function (req, res, next) {
+  try {
+    const todos = await TodoModel.find();
+    console.log("todos", todos);
+    res.json({ result: todos });
+  } catch (error) {
+    console.error("Error in Get Todos/", error);
+    next(error);
+  }
+});
+
+router.get("/:id", async function (req, res, next) {
+  try {
+    const params = req.params;
+    const id = params.id;
+    if (!validateObjectId(id)) {
+      return res.status(400).json({ message: "Id must be ObjectId type" });
+    }
+    const todoById = await TodoModel.findById(id);
+    if (!todoById) {
+      const notFound = `Todo with ${id} is not found`;
+      return res.status(404).json({ result: null, message: notFound });
+    } else {
+      res.json({ result: todoById });
+    }
+  } catch (error) {
+    console.error("Error in Get Todos/:id", error);
+    next(error);
+  }
 });
 
 router.post("/", upload.none(), async function (req, res, next) {
@@ -22,8 +51,9 @@ router.post("/", upload.none(), async function (req, res, next) {
     });
     console.log("mock todo model from backend", todo);
     await todo.save();
-    res.send({ result: "Create todo success" });
+    res.json({ message: "Create todo successfully" });
   } catch (error) {
+    console.error("Error in Post Todos/", error);
     next(error);
   }
 });
@@ -32,6 +62,9 @@ router.put("/", upload.none(), async function (req, res, next) {
   try {
     const body = req.body;
     const id = body.id;
+    if (!validateObjectId(id)) {
+      return res.status(400).json({ message: "Id must be ObjectId type" });
+    }
     const filter = { _id: id };
     const replaceTodo = await TodoModel.findByIdAndUpdate(
       filter,
@@ -46,13 +79,14 @@ router.put("/", upload.none(), async function (req, res, next) {
       }
     );
     if (!replaceTodo) {
-      return res.status(404).send({ message: `Todo with id ${id} not found` });
+      return res.status(404).json({ message: `Todo with id ${id} not found` });
     }
     res.send({
       message: "Successfully updated",
       result: replaceTodo.toObject(),
     });
   } catch (error) {
+    console.error("Error in Put Todos/", error);
     next(error);
   }
 });
@@ -62,12 +96,16 @@ router.delete("/:id", async function (req, res, next) {
     const params = req.params;
     const id = params.id;
     console.log("id to delete from params", id);
+    if (!validateObjectId(id)) {
+      return res.status(400).json({ message: "Id must be ObjectId type" });
+    }
     const deleteTodo = await TodoModel.findByIdAndDelete(id);
     if (!deleteTodo) {
-      return res.status(404).send({ message: `Todo with ${id} not found` });
+      return res.status(404).send({ message: `Todo #${id} not found` });
     }
     res.send({ message: "Successfully deleted" });
   } catch (error) {
+    console.error("Error in Delete Todos/:id");
     next(error);
   }
 });
